+++ src/main/scala/ca/ferlab/ferload/client/clients/FerloadClient.scala
@@ -3,25 +3,19 @@ package ca.ferlab.ferload.client.clients
 import ca.ferlab.ferload.client.clients.inf.IFerload
 import ca.ferlab.ferload.client.configurations.{FerloadUrl, UserConfig}
 import org.apache.commons.lang3.StringUtils
-import org.apache.http.client.methods.{HttpGet, HttpPost, HttpRequestBase}
+import org.apache.http.HttpHeaders
+import org.apache.http.client.methods.{HttpGet, HttpPost}
 import org.apache.http.entity.{ContentType, StringEntity}
-import org.apache.http.impl.client.{CloseableHttpClient, HttpClientBuilder}
-import org.apache.http.util.EntityUtils
-import org.apache.http.{HttpHeaders, HttpResponse}
 import org.json.JSONObject
 
 import java.io.File
 import java.net.URL
 import scala.io.Source
-import scala.jdk.CollectionConverters.MapHasAsScala
 
-class FerloadClient(userConfig: UserConfig) extends IFerload {
+class FerloadClient(userConfig: UserConfig) extends BaseHttpClient with IFerload {
 
-  val httpBuilder: HttpClientBuilder = HttpClientBuilder.create()
-  val http: CloseableHttpClient = httpBuilder.build()
   // lazy because config may be not available
   lazy val url = new URL(userConfig.get(FerloadUrl))
-  private val charset = "UTF-8"
   private val separator = "\n"
 
   override def getDownloadLinks(token: String, manifest: File): Map[String, String] = {
@@ -48,22 +42,6 @@ class FerloadClient(userConfig: UserConfig) extends IFerload {
     val (body, status) = executeHttpRequest(httpRequest)
     if (status != 200) throw new IllegalStateException(formatExceptionMessage("Failed to retrieve Ferload config", status, body))
     body.map(new JSONObject(_)).get // throw exception if null, that's ok
-  }
-
-  private def executeHttpRequest(request: HttpRequestBase): (Option[String], Int) = {
-    val response: HttpResponse = http.execute(request)
-    val body = Option(response.getEntity).map(e => EntityUtils.toString(e, charset))
-    // always properly close
-    EntityUtils.consumeQuietly(response.getEntity)
-    (body, response.getStatusLine.getStatusCode)
-  }
-
-  private def formatExceptionMessage(message: String, status: Int, body: Option[String]) = {
-    s"$message, code: $status, message:\n${body.getOrElse("")}"
-  }
-
-  private def toMap(body: Option[String]): Map[String, String] = {
-    body.map(new JSONObject(_).toMap.asScala.map({ case (key, value) => (key, value.toString) }).toMap).getOrElse(Map())
   }
 
 }
